Initial Commit
Got a list of connections and sent them data. I have not implemented the receive yet for the data. I was using that netstat to look a valid connections and got confused. When a new ESP comes online it lists it, but it deletes the old. I am wondering if the connection is still established.